requirements.txt
Django>=5.0,<6.1
djangorestframework
django-cors-headers
channels
daphne
channels_redis
dj-database-url
psycopg2-binary
